requirements.txt
alembic==1.5.7
click==7.1.2
fake-useragent==0.1.11
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
python-dotenv==0.15.0
selenium==3.141.0
SQLAlchemy==1.3.23
pytest==6.2.2
pytest-mock==3.5.1
